+++ src/main/java/com/plivo/api/PlivoAPIService.java
@@ -15,6 +15,9 @@ import com.plivo.api.models.conference.*;
 import com.plivo.api.models.endpoint.*;
 import com.plivo.api.models.enduser.*;
 import com.plivo.api.models.identity.*;
+import com.plivo.api.models.maskingsession.MaskingSession;
+import com.plivo.api.models.maskingsession.MaskingSessionUpdateResponse;
+import com.plivo.api.models.maskingsession.MaskingSessionUpdater;
 import com.plivo.api.models.media.Media;
 import com.plivo.api.models.media.MediaResponse;
 import com.plivo.api.models.message.Message;
@@ -350,6 +353,23 @@ public interface PlivoAPIService {
   @POST("Account/{authId}/JWT/Token/")
   Call<TokenCreateResponse> tokenCreate(@Path("authId") String authId,
                                                      @Body TokenCreator tokenCreator);
+
+  @GET("Account/{authId}/Masking/Session")
+  Call<ListResponse<MaskingSession>> maskingSessionList(@Path("authId") String authId,
+                                                        @QueryMap Map<String, Object> params);
+
+  @GET("Account/{authId}/Masking/Session/{sessionUuid}")
+  Call<MaskingSession> maskingSessionGet(@Path("authId") String authId, @Path("sessionUuid") String sessionUuid);
+
+  @DELETE("Account/{authId}/Masking/Session/{sessionUuid}/")
+  Call<ResponseBody> maskingSessionDelete(@Path("authId") String authId,
+                                    @Path("sessionUuid") String sessionUuid);
+
+  @POST("Account/{authId}/Endpoint/{endpointId}/")
+  Call<MaskingSessionUpdateResponse> maskingSessionUpdate(@Path("authId") String authId,
+                                                          @Path("sessionUuid") String sessionUuid, @Body
+                                                          MaskingSessionUpdater maskingSessionUpdater);
+
   @GET("Account/{authId}/Endpoint/")
   Call<ListResponse<Endpoint>> endpointList(@Path("authId") String authId,
                                             @QueryMap Map<String, Object> params);