+++ src/main/java/com/plivo/api/models/maskingsession/MaskingSessionCreateResponse.java
@@ -1,0 +1,6 @@
+package com.plivo.api.models.maskingsession;
+
+import com.plivo.api.models.base.BaseResponse;
+
+public class MaskingSessionCreateResponse extends BaseResponse {
+}